Decompose engine.py
