+++ sdks/python/src/opik/evaluation/engine/types.py
@@ -1,0 +1,7 @@
+from .. import test_result
+from typing import Protocol
+
+
+class EvaluationTask(Protocol):
+    def __call__(self) -> test_result.TestResult:
+        pass
